refactor: performance의 type 필드 enum으로 변경
