+++ src/main/java/com/wanted/preonboarding/ticket/domain/entity/Performance.java
@@ -22,21 +22,28 @@ public class Performance extends AuditInformation {
     @GenericGenerator(name = "uuid2", strategy = "uuid2")
     @Column(columnDefinition = "BINARY(16)")
     private UUID id;
+
     @Column(nullable = false)
     @Comment("공연/전시 이름")
     private String name;
+
     @Column(nullable = false)
     @Comment("가격")
     private int price;
+
     @Column(nullable = false)
     @Comment("회차")
     private int round;
+
+    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     @Comment("NONE, CONCERT, EXHIBITION")
-    private int type;
+    private PerformanceType type;
+
     @Column(nullable = false)
     @Comment("공연 일시")
     private LocalDateTime start_date;
+
     @Column(nullable = false, name = "is_reserve", columnDefinition = "varchar default 'disable'")
     @Comment("예약 여부")
     private String isReserve;
